Move quizData JSON into quizData folder
