+++ pages/api/studentPortal/quizzes/quizData/careerQuiz.ts
