+++ pages/resources/quizzes/firstProgrammingLanguage/index.tsx
@@ -5,16 +5,16 @@ import BluePrint from "../../../../components/resources/quizzes/shared/BluePrint
 import SkillSelections from "../../../../components/resources/quizzes/shared/SkillSelections";
 import StartQuiz from "../../../../components/resources/quizzes/shared/StartQuiz";
 import {
-  QuizOptionViewState,
-  QuizViewState,
-  UserInput,
+    QuizOptionViewState,
+    QuizViewState,
+    UserInput,
 } from "../../../../components/resources/quizzes/shared/types";
 import QuizTransition from "../../../../components/ui/animations/QuizTransition";
 import { INSERT_CODING_LANGUAGE_QUIZ_RESPONSE } from "../../../../graphql/quizzes/insertCodingLanguageQuiz";
 import { UPDATE_CODING_LANGUAGE_QUIZ_RESPONSE } from "../../../../graphql/quizzes/updateCodingLanguageQuiz";
 import { computeLanguageScore } from "../../../api/studentPortal/quizzes/firstProgrammingLanguage/computeScore";
-import { quizData } from "../../../api/studentPortal/quizzes/firstProgrammingLanguage/firstProgrammingLanguage";
 import { getPreferredLanguageForQuizResults } from "../../../api/studentPortal/quizzes/firstProgrammingLanguage/getPreferredLanguage";
+import { quizData } from "../../../api/studentPortal/quizzes/quizData/firstProgrammingLanguage";
 
 export enum Stage {
   START,
